Change swap to use generate SE function
Currently the `do_swap` function manually generates the SE address
for the swap but does not add it to `se_backup_keys` nor assign a
`tx_backup_addr` to it.

Apparently there is no need to re-implement the SE address generation
logic as there is a `get_new_state_entity_address` function that already
does that.