Instead of using `tx_backup_addr`, use `proof_key` to set the new owner
The `swap/second` returns a SCEAddress with null `tx_backup_addr` and
`transfer_sender` function uses this to replace the previous owner.
If the `tx_backup_addr` is null, the previous owner persists and then
the error `Backup Tx Receiving Address not found in this wallet` occurs
in the swap process.
Using `proof_key` to set the new owner in the backup transaction fix this.